ODBC-313 Partial fix and relevant testcases
Added session tracking use to avoid extra queries to server. Removed
mysql_sttm_reset calls. Qureies, initiated by the connector, assembled
into a batch if possible(it;s mostly about queries sent at connect time).

+++ ma_typeconv.c
@@ -237,7 +237,7 @@ size_t MADB_ConvertNumericToChar(SQL_NUMERIC_STRUCT *Numeric, char *Buffer, int 
     }
 
     /* If scale >= precision, we still can have no truncation */
-    if (Length > Numeric->precision + 1/*dot*/ && Scale < Numeric->precision)
+    if (Length > (unsigned int)(Numeric->precision + 1)/*dot*/ && Scale < Numeric->precision)
     {
       *ErrorCode= MADB_ERR_01S07;
       Length = Numeric->precision + 1/*dot*/;
